Moved to src and added first test
Note that this removes the Aqua tests!
