+++ test/runtests.jl
@@ -47,4 +47,4 @@ end
   include("decasymbolic.jl")
 end
 
-include("aqua.jl")
+# include("aqua.jl")
